1thyroid nodule segmentation/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
Augmentor==0.2.12
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cycler==0.11.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.4.3
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0